\name{FHclust-package}
\alias{FHclust-package}
\alias{FHclust}
\docType{package}
\title{
\packageTitle{FHclust}
}
\description{
\packageDescription{FHclust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FHclust}
\packageIndices{FHclust}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FHclust}

Maintainer: \packageMaintainer{FHclust}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
